refactor: refactored jamba  model specific functions to a separate folder
