+++ src/utils/aws-bedrock/response-converter.ts
@@ -3,6 +3,7 @@ import { convertClaudeResponse } from './claude/convert-claude-response'
 import { convertLlamaResponse } from './llama/convert-llama-response'
 import { convertNovaResponse } from './nova/convert-nova-response'
 import { convertCohereResponse } from './cohere/convert-cohere-response'
+import { convertJambaResponse } from './jamba/convert-jamba-response'
 
 function convertMistralResponse(responseBody: any): Partial<StandardResponse> {
   const generatedText = responseBody.outputs?.[0]?.text || ''
@@ -33,45 +34,6 @@ function convertMistralResponse(responseBody: any): Partial<StandardResponse> {
       inputTokens: 0,
       outputTokens: 0,
       totalTokens: 0
-    }
-  }
-}
-
-function convertJambaResponse(responseBody: any): Partial<StandardResponse> {
-  const choice = responseBody.choices?.[0]
-  const message = choice?.message
-
-  const content: any[] = []
-
-  if (message?.content) {
-    content.push({ type: 'text' as const, text: message.content })
-  }
-
-  // Handle tool calls if present
-  if (message?.tool_calls) {
-    message.tool_calls.forEach((call: any) => {
-      content.push({
-        type: 'tool-call' as const,
-        toolCallId: call.id,
-        toolName: call.function.name,
-        args: typeof call.function.arguments === 'string'
-          ? JSON.parse(call.function.arguments)
-          : call.function.arguments
-      })
-    })
-  }
-
-  const finishReason = choice?.finish_reason === 'stop' ? 'stop'
-    : choice?.finish_reason === 'length' ? 'length'
-      : choice?.finish_reason === 'tool_calls' ? 'tool-calls' : 'stop'
-
-  return {
-    content,
-    finishReason,
-    usage: {
-      inputTokens: responseBody.usage?.prompt_tokens || 0,
-      outputTokens: responseBody.usage?.completion_tokens || 0,
-      totalTokens: responseBody.usage?.total_tokens || 0
     }
   }
 }
